Fix race conditions in tests
The test originally relied upon a `die` `MVar` to kill the `server`
thread. This lead to a race condition where the server could still be
running when another server was attempting to start. This new
configuration allows the server to signal when it is ready to receive
requests and it races with the action itself. This allows the `race_` to
complete when the server has been fully killed by the RTS.